test: added test for all methods from OCOAuthRepository
